+++ src/bin/hnefatafl-server-full/tests.rs
@@ -350,7 +350,6 @@ fn change_password() -> anyhow::Result<()> {
 }
 
 #[test]
-#[allow(clippy::cast_possible_truncation)]
 #[allow(clippy::float_cmp)]
 fn check_update_rd() -> anyhow::Result<()> {
     let mut server = ServerFull {
@@ -365,9 +364,7 @@ fn check_update_rd() -> anyhow::Result<()> {
     }
 
     assert!(!server.check_update_rd());
-    server.ran_update_rd = UnixTimestamp(
-        (Timestamp::now() - (TWO_MONTHS.round_ties_even() as i64).seconds()).as_second(),
-    );
+    server.ran_update_rd.0 = Timestamp::now() - TWO_MONTHS_MICRO_SECONDS.microseconds();
     assert!(server.check_update_rd());
 
     if let Some(account) = server.accounts.0.get_mut("david") {